instant upgrade, to Inferno utils level

+++ src/cmd/ld/Plan9.c
@@ -1,7 +1,7 @@
 #include	"l.h"
 
 void*
-mysbrk(ulong size)
+mysbrk(usize size)
 {
 	return sbrk(size);
 }
